gs: Split driver into multiple files
